requirement.txt

# Linear Algebra Agent

streamlit
google-genai
python-dotenv
